requirements.txt
fastapi
uvicorn[standard]
langgraph
langchain-google-genai
tavily-python
python-dotenv